fix(wallet): remove NFT collection name and panel header overlap
