+++ components/brave_wallet_ui/components/desktop/views/crypto/index.tsx
@@ -4,20 +4,13 @@
 // you can obtain one at https://mozilla.org/MPL/2.0/.
 
 import * as React from 'react'
-import {
-  Route,
-  useHistory,
-  Switch,
-  Redirect
-} from 'react-router-dom'
+import { Route, useHistory, Switch, Redirect } from 'react-router-dom'
 import { useSelector } from 'react-redux'
 
 // utils
 import { loadTimeData } from '../../../../../common/loadTimeData'
 import { getLocale } from '../../../../../common/locale'
-import {
-  useSafeUISelector
-} from '../../../../common/hooks/use-safe-selector'
+import { useSafeUISelector } from '../../../../common/hooks/use-safe-selector'
 import { UISelectors } from '../../../../common/selectors'
 import { openWalletSettings } from '../../../../utils/routes-utils'
 import {
@@ -35,6 +28,12 @@ import {
 
 // hooks
 import { useQuery } from '../../../../common/hooks/use-query'
+import {
+  usePortfolioVisibleNetworks //
+} from '../../../../common/hooks/use_portfolio_networks'
+import {
+  usePortfolioAccounts //
+} from '../../../../common/hooks/use_portfolio_accounts'
 
 // style
 import { StyledWrapper } from './style'
@@ -68,6 +67,7 @@ import { PageTitleHeader } from '../../card-headers/page-title-header'
 import { MarketAsset } from '../market/market_asset'
 import { ExploreWeb3View } from '../explore_web3/explore_web3'
 import { DappDetails } from '../explore_web3/web3_dapp_details'
+import { NftCollection } from '../nfts/components/nft_collection'
 
 export interface Props {
   sessionRoute: string | undefined
@@ -81,6 +81,10 @@ export const CryptoView = ({ sessionRoute }: Props) => {
   const { accountToRemove, showAccountModal, selectedAccount } = useSelector(
     ({ accountsTab }: { accountsTab: AccountsTabState }) => accountsTab
   )
+
+  // custom hooks
+  const { visiblePortfolioNetworks } = usePortfolioVisibleNetworks()
+  const { usersFilteredAccounts } = usePortfolioAccounts()
 
   // queries
   const {
@@ -256,6 +260,16 @@ export const CryptoView = ({ sessionRoute }: Props) => {
           exact
         >
           <PortfolioFungibleAsset />
+        </Route>
+
+        <Route
+          path={WalletRoutes.PortfolioNFTCollection}
+          exact
+        >
+          <NftCollection
+            networks={visiblePortfolioNetworks}
+            accounts={usersFilteredAccounts}
+          />
         </Route>
 
         <Route path={WalletRoutes.Portfolio}>
